+++ ExtendedDateTimeFormat/Extensions/DateTimeOffsetExtensions.cs
@@ -4,7 +4,7 @@
     {
         public static ExtendedDateTime ToExtendedDateTime(this DateTimeOffset d)
         {
-            return new ExtendedDateTime(d.Year, (byte)d.Month, (byte)d.Day, (byte)d.Hour, (byte)d.Minute, (byte)d.Second, d.Offset);
+            return new ExtendedDateTime(d.Year, d.Month, d.Day, d.Hour, d.Minute, d.Second, d.Offset);
         }
     }
 }
